replace metric, rename the metricrepoter to metric
